requirements.txt
discord.py~=2.1.0
requests~=2.28.1
bs4~=0.0.1
beautifulsoup4~=4.11.1
SQLAlchemy~=1.4.44
MySQL==0.0.3